feat: adding abstraction and granularity to the values we seraliaze / compare too
