+++ beams/tests/artifacts/egg_generator.py
@@ -9,7 +9,7 @@ import py_trees
 from apischema import serialize
 
 from beams.tree_config.action import IncPVActionItem, SetPVActionItem
-from beams.tree_config.base import BehaviorTreeItem, EPICSValue, FixedValue
+from beams.tree_config.base import BehaviorTreeItem
 from beams.tree_config.composite import SequenceItem
 from beams.tree_config.condition import (BinaryConditionItem,
                                          BoundedConditionItem,
@@ -17,6 +17,7 @@ from beams.tree_config.condition import (BinaryConditionItem,
 from beams.tree_config.idiom import CheckAndDoItem
 from beams.tree_config.py_trees import (RunningItem, StatusQueueItem,
                                         SuccessItem)
+from beams.tree_config.value import EPICSValue, FixedValue
 
 
 # egg 1
